s2: camera: fix compilation errors

+++ camera/QCamera2/HAL/QCameraMuxer.cpp
@@ -2347,7 +2347,7 @@ void QCameraMuxer::composeMpo(cam_compose_jpeg_info_t* main_Jpeg,
                     mpo_compose_info.primary_image.buf_filled_len);
             fchmod(file_fd_main, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
             CDBG("%s: written number of bytes for main Image %ld\n",
-                    __func__, written_len);
+                    __func__, (long)written_len);
             close(file_fd_main);
         }
 
@@ -2364,7 +2364,7 @@ void QCameraMuxer::composeMpo(cam_compose_jpeg_info_t* main_Jpeg,
                     mpo_compose_info.aux_images[0].buf_filled_len);
             fchmod(file_fd_aux, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
             CDBG("%s: written number of bytes for Aux Image %ld\n",
-                    __func__, written_len);
+                    __func__, (long)written_len);
             close(file_fd_aux);
         }
     }
@@ -2392,7 +2392,7 @@ void QCameraMuxer::composeMpo(cam_compose_jpeg_info_t* main_Jpeg,
                     m_pRelCamMpoJpeg->size);
             fchmod(file_fd_mpo, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
             CDBG("%s: written number of bytes for MPO Image %ld\n",
-                    __func__, written_len);
+                    __func__, (long)written_len);
             close(file_fd_mpo);
         }
     }
